Simplified structure of `curry` and made it faster.
Lesson learned: `<object>NULL is not None` (and `NULL is not <PyObject*>None`).
When a C API function says it can receive an object of a certain type or NULL,
it means NULL (not `None`) even if the binding is done as an `object`.
This came up regarding `PyObject_Call`, which I initially had, but needed to
change to something slower in order to pass tests.