Add `interleave`, which is about 5x faster than in `toolz`.
Lesson learned: when using `PyObject* PyIter_Next(object o)`, you need to
`Py_DECREF` the value when you are finished with it (such as after you
cast it to <object>).  The rationale, I suppose, is that `PyIter_Next` is
commonly used to construct lists and tuples, which steal references.
In other words, when constructing lists and tuples with `PyIter_Next`,
you don't need to call either `Py_DECREF` or `Py_INCREF`.