Issue #5: add `cytoolz.utils.include_dirs` to allow other project to use CyToolz' C API.
A minimal example is included in the docstring, although a basic example
should also be in the hosted documentation (whenever it exists).

This should be regarded as an advanced feature for advanced users, so
typical users should not be encouraged to use it.  This is the main
reason `include_dirs` is tucked away in `cytoolz.utils`.