Add tests to ensure None is safe input.  Addresses #6.
Arguments that are strongly typed Python object (like `dict`) may also be None.
Although many C API functions check for type, some don't, and passing None may
result in an uncatchable SystemError or possibly a segfault on some systems.
This actually turned out to be a very minor issue, but is one we should
continue to test for.