Rename additional C bindings defined in `cytoolz.cpython`
This differentiates the additional bindings with a common convention.
Currently, we only change bindings to return pointers instead of Python
objects, and we change the name from "Py*" to "Ptr*".  I expect we will
add more bindings in the future, including passing pointers to functions.
In this case, the name should be something like "*_ptr".